Implements a fix for stereo correction with TS geometries. Not carefully tested
